resume and popup
Experimenting with creating a popup to let folks know that my page is a little dated, but resume is recent.  This is my attempt to "brush the dust off" my dev skills.

Oh and of course I found things I didn't like on my first resume commit, so... *hits redo*